+++ GUI/src/pages/AdvisorsPage.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import OptionsPanel, { Option } from '../components/MetricAndPeriodOptions';
 import MetricsCharts from '../components/MetricsCharts';
 import { MetricOptionsState } from '../components/MetricAndPeriodOptions/types';
-import { chartDataKey, formatDate, translateChartKeys } from '../util/charts-utils';
+import { chartDataKey, formatDate, getAdvisorsList, getChartResponse, translateChartKeys } from '../util/charts-utils';
 import { debounceTime, distinctUntilChanged, switchMap } from 'rxjs/operators';
 import { Subject } from 'rxjs';
 import {
@@ -235,16 +235,7 @@ const AdvisorsPage: React.FC = () => {
 
       const res = result.response;
 
-      const advisorsList = Array.from(new Set(res.map((advisor: any) => advisor.customerSupportId)))
-        .map((id: any) => res.find((e: any) => e.customerSupportId == id))
-        .map((e) => {
-          return {
-            id: e?.customerSupportId ?? '',
-            labelKey: e?.customerSupportFullName ?? '',
-            color: randomColor(),
-            isSelected: true,
-          };
-        });
+      const advisorsList = getAdvisorsList(res);
 
       if (advisorsList.length > advisors.current.length) {
         const updatedMetrics = [...advisorsMetrics];
@@ -260,37 +251,8 @@ const AdvisorsPage: React.FC = () => {
         setAdvisorsList([]);
       }
 
-      const response = res
-        .flat(1)
-        .map((entry: any) => ({
-          ...translateChartKeys(entry, chartDataKey),
-          [chartDataKey]: new Date(entry[chartDataKey]).getTime(),
-        }))
-        .reduce((a: any, b: any) => {
-          const dateRow = a.find((i: any) => i[chartDataKey] === b[chartDataKey]);
-          if (dateRow) {
-            dateRow[b[t('chart.customerSupportFullName')]] = b[t('chart.count')];
-          } else {
-            a.push({
-              [chartDataKey]: b[chartDataKey],
-              [b[t('chart.customerSupportFullName')]]: b[t('chart.count')],
-            });
-          }
-          return a;
-        }, []);
-
-      const chartResponse = response.map((e: any) => {
-        const res = { ...e };
-        advisorsList.forEach((i) => {
-          if (!(i.labelKey in e)) {
-            res[i.labelKey] = 0;
-          }
-        });
-        return res;
-      });
-
       chartData = {
-        chartData: chartResponse,
+        chartData: getChartResponse(res, advisorsList),
         colors: advisorsMetrics[3].subOptions!.map(({ labelKey, color }) => {
           return {
             id: labelKey,
