+++ GUI/src/types/advisor.ts
@@ -1,0 +1,6 @@
+export interface Advisor {
+  id: string;
+  labelKey: string;
+  color: string;
+  isSelected: boolean;
+}
